@@ -1,6 +1,6 @@
 module git.sr.ht/~spc/mqttcli
 
-go 1.15
+go 1.16
 
 require (
 	github.com/eclipse/paho.mqtt.golang v1.3.5
